backend/requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9.0
python-decouple>=3.8
gunicorn>=21.0.0
whitenoise>=6.5.0
django-cors-headers>=4.3.1
razorpay>=1.4.0
geopy>=2.3.0
